Unit-tests were added
