+++ DEV-7/DetermineTriangleType.Tests/packages.config
