+++ DEV-7/DetermineTriangleType/Versatile.cs
@@ -1,6 +1,6 @@
 ﻿namespace DetermineTriangleType
 {
-    class Versatile:Triangle
+    public class Versatile:Triangle
     {
         const string VERSALITE = "Triangle is versalite";
         public Versatile(Sides sides) : base(sides) { }
